Refactor mic permission
